requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
boto3==1.14.23
botocore==1.17.23
bs4==0.0.1
Django==3.0.7
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-grappelli==2.14.2
django-js-asset==1.2.2
django-storages==1.9.1
django-taggit==1.3.0
django-user-agents==0.4.0
docutils==0.15.2
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
whitenoise==5.1.0
